Kernel: Move PCI PIIX4 IDE controller code to an appropriate directory
