+++ Meta/gn/secondary/Kernel/BUILD.gn
@@ -328,7 +328,6 @@ source_set("arch_sources") {
       "Arch/x86_64/Interrupts/IOAPIC.cpp",
       "Arch/x86_64/Interrupts/PIC.cpp",
       "Arch/x86_64/PCI/Controller/HostBridge.cpp",
-      "Arch/x86_64/PCI/IDELegacyModeController.cpp",
       "Arch/x86_64/PCI/Initializer.cpp",
       "Arch/x86_64/PCI/MSI.cpp",
       "Arch/x86_64/PCSpeaker.cpp",
@@ -565,6 +564,7 @@ executable("Kernel_bin") {
     "Devices/Storage/ATA/ATAPort.cpp",
     "Devices/Storage/ATA/GenericIDE/Channel.cpp",
     "Devices/Storage/ATA/GenericIDE/Controller.cpp",
+    "Devices/Storage/ATA/PIIX4IDE/Controller.cpp",
     "Devices/Storage/DiskPartition.cpp",
     "Devices/Storage/NVMe/NVMeController.cpp",
     "Devices/Storage/NVMe/NVMeInterruptQueue.cpp",
